Refactor debug UI into separate component
