+++ src/UI/IUIElement.cs
@@ -1,0 +1,6 @@
+namespace RenderMaster;
+
+public interface IUIElement
+{
+    void AfterBegin();
+}
